Dude&Poo-float changes

+++ Engine/Dude.h
@@ -9,14 +9,14 @@ public:
 	void ClampToScreen();
 	void Draw( Graphics& gfx ) const;
 	void Update( const Keyboard& kbd );
-	int GetX() const;
-	int GetY() const;
+	float GetX() const;
+	float GetY() const;
 	int GetWidth() const;
 	int GetHeight() const;
 private:
-	int x = 400;
-	int y = 300;
-	static constexpr int speed = 1;
+	float x = 400;
+	float y = 300;
+	static constexpr float speed = 1;
 	static constexpr int width = 20;
 	static constexpr int height = 20;
 };
